Add files via upload
The goal of this project is to develop a Sinirio, a web-based application for creating and managing online surveys. Sinirio is designed to help professionals and researchers to quickly and easily create and manage their surveys. It provides an intuitive interface that allows users to easily create survey questions, assign them to different groups, and track responses.